requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
plotly
mysql-connector-python